extract common code and handle promise rejections in worker hook
